backend/requirements.txt
Package       Version
------------- -------
click         8.1.7
colorama      0.4.6
Flask         2.2.5
itsdangerous  2.0.1
Jinja2        3.1.3
MarkupSafe    2.1.3
pip           23.3.1
python-dotenv 0.21.0
setuptools    68.2.2
Werkzeug      2.3.8
wheel         0.41.2
